Refactor: extracted DamageDetector
