+++ ios/FrameExtractor.swift
@@ -10,28 +10,6 @@ import CoreML
 import Vision
 import ImageIO
 
-extension CGImagePropertyOrientation {
-  /**
-   Converts a `UIImageOrientation` to a corresponding
-   `CGImagePropertyOrientation`. The cases for each
-   orientation are represented by different raw values.
-   
-   - Tag: ConvertOrientation
-   */
-  init(_ orientation: UIImage.Orientation) {
-    switch orientation {
-    case .up: self = .up
-    case .upMirrored: self = .upMirrored
-    case .down: self = .down
-    case .downMirrored: self = .downMirrored
-    case .left: self = .left
-    case .leftMirrored: self = .leftMirrored
-    case .right: self = .right
-    case .rightMirrored: self = .rightMirrored
-    }
-  }
-}
-
 class FrameExtractor : NSObject, AVCaptureVideoDataOutputSampleBufferDelegate {
   private let captureSession = AVCaptureSession()
   private let sessionQueue = DispatchQueue(label: "session queue")
@@ -39,16 +17,13 @@ class FrameExtractor : NSObject, AVCaptureVideoDataOutputSampleBufferDelegate {
   private let position = AVCaptureDevice.Position.back
   private let quality = AVCaptureSession.Preset.medium
   private let context = CIContext()
-  var damageDetected: ((MLMultiArray?) -> Void)?
   var frameCaptured: ((UIImage?) -> Void)?
-  private var throttler: Throttler!
   
   
   override init() {
     print("init")
     super.init()
     checkPermission()
-    throttler = Throttler(seconds: 0.25)
     sessionQueue.async { [unowned self] in // sessionQueue might be suspended after checkPermission. unowned self prevents a possible retain cycle.
       self.configureSession()
       self.captureSession.startRunning() // Session is started on the sessionQueue, frames are processed on a separate queue
@@ -86,7 +61,7 @@ class FrameExtractor : NSObject, AVCaptureVideoDataOutputSampleBufferDelegate {
     guard captureSession.canAddInput(captureDeviceInput) else { return }
     captureSession.addInput(captureDeviceInput)
     let videoOutput = AVCaptureVideoDataOutput()
-    videoOutput.setSampleBufferDelegate(self, queue: DispatchQueue(label: "sample buffer")) // Self is an AVCaptureVideoDataOutputSampleBufferDelegate, frames are processed on a different queue to avoid pile up
+    videoOutput.setSampleBufferDelegate(self, queue: DispatchQueue(label: "sample buffer"))
     guard captureSession.canAddOutput(videoOutput) else { return }
     captureSession.addOutput(videoOutput)
     guard let connection = videoOutput.connection(with: AVFoundation.AVMediaType.video) else { return }
@@ -96,59 +71,18 @@ class FrameExtractor : NSObject, AVCaptureVideoDataOutputSampleBufferDelegate {
     connection.isVideoMirrored = position == .front
   }
   
+  private func imageFromSampleBuffer(sampleBuffer: CMSampleBuffer) -> UIImage? {
+    guard let imageBuffer = CMSampleBufferGetImageBuffer(sampleBuffer) else { return nil }
+    let ciImage = CIImage(cvPixelBuffer: imageBuffer)
+    guard let cgImage = context.createCGImage(ciImage, from: ciImage.extent) else { return nil }
+    return UIImage(cgImage: cgImage)
+  }
   
   func captureOutput(_ output: AVCaptureOutput, didOutput sampleBuffer: CMSampleBuffer, from connection: AVCaptureConnection) {
-    guard let imageBuffer = CMSampleBufferGetImageBuffer(sampleBuffer) else { return }
-    let ciImage = CIImage(cvPixelBuffer: imageBuffer)
-    guard let cgImage = context.createCGImage(ciImage, from: ciImage.extent) else { fatalError("Unable to create \(CGImage.self) from \(ciImage).")  }
-    let uiImage = UIImage(cgImage: cgImage)
-    let orientation = CGImagePropertyOrientation(uiImage.imageOrientation)
-
-    throttler.throttle(block: { [unowned self] in
-      let handler = VNImageRequestHandler(ciImage: ciImage, orientation: orientation)
-      do {
-        try handler.perform([self.classificationRequest])
-      } catch {
-        print("Failed to perform classification.\n\(error.localizedDescription)")
-      }
-    }, queue: DispatchQueue.global(qos: .userInitiated))
+    guard let uiImage = imageFromSampleBuffer(sampleBuffer: sampleBuffer) else { return }
     
     DispatchQueue.main.async { [unowned self] in // Image is sent to the delegate on the main thread.
       self.frameCaptured?(uiImage)
     }
   }
-  
-  /// Updates the UI with the results of the classification.
-  /// - Tag: ProcessClassifications
-  func processClassifications(for request: VNRequest, error: Error?) {
-    DispatchQueue.main.async { [unowned self] in
-      guard let results = request.results else {
-        print("Unable to classify image.\n\(error!.localizedDescription)")
-        return
-      }
-      
-      let classifications = results as! [VNCoreMLFeatureValueObservation]
-      
-      let obs : VNCoreMLFeatureValueObservation = (classifications.first)!
-      let m: MLMultiArray = obs.featureValue.multiArrayValue!
-      
-      self.damageDetected?(m)
-    }
-  }
-  
-  lazy var classificationRequest: VNCoreMLRequest = {
-    do {
-      let model = try VNCoreMLModel(for: RoadDamageModel().model)
-      
-      let request = VNCoreMLRequest(model: model, completionHandler: { [weak self] request, error in
-        self?.processClassifications(for: request, error: error)
-      })
-      
-      request.imageCropAndScaleOption = .centerCrop
-      
-      return request
-    } catch {
-      fatalError("Failed to load Vision ML model: \(error)")
-    }
-  }()
 }
